Implement Ability to Die
Add key adventure game feature: the ability to die. Recovering from
a death will take the git command feature set. For future reference
the act of dying in the game mimicks reaching a spot in your project
when you have made a discreet set of changes that should be committed