Add movement method for Character Class
Add a wrapper method for characters so they can move. Have
the method call be a wrapper for the character class so that
many characters can be on the board whilst all calling on one
map object to determine where they move.

Note: in order for the tests to work, the recordable class
must be comparable, so define __eq__ and __ne__ methods to
see if the program is writing files correctly.