Add cli for issuing game play instructions
Substaniate a CLI framework the allows the player to move
about / affect the environment via a command line interface.
Test to ensure player can properly move with this interface.

To Do: Prevent player from moving off the map. All, a display
of location and what can be done with it is also needed.