Sync Mock-Data Folder with Its Subrepo
Sync the saved state of mock-data is the main repo with the saved
state of mock-data in the sub repo. As tests consistently resets
the sub-repo to a committed state. This is the only way to make
sure running the unittests doesn't show up as a change to the
repo.