Implement Trapped Room
Implement a trap, including the check for traps. Traps should have
the following effects: kill the character, cause a status effect,
record change of player state, and an ending of the player command
loop (if in effect ending the game).

TODO: write the game as two loops: an active loop when the player
is alive that has access to all commands, and a passive loop when
the player is dead that has access to the git commands only.