Change constructor for mock object tests
Alter the program so that a test can be run on the system's
ability to produce a commit. For that test to be run, a
GitCmd class (or subclass) must call it's do_commit method.
In order to clean up from that test, the test repo must be
reset to HEAD~1, rather than have HEAD forcibly checked out,
which would leave the test repo in detached HEAD state.